requirements.txt

diffusers==0.26.3
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
torch==2.2.1
torchvision==0.17.1